Cleanup based on hints

+++ src/main/java/org/comdnmr/fit/calc/CurveFit.java
@@ -41,8 +41,8 @@ public class CurveFit {
     public List<ParValueInterface> getParValues() {
         List<ParValueInterface> dataValues = new ArrayList<>();
         parMap.keySet().stream().sorted().filter((parName) -> (parMap.containsKey(parName + ".sd"))).forEachOrdered((parName) -> {
-            double value = parMap.get(parName).doubleValue();
-            double err = parMap.get(parName + ".sd").doubleValue();
+            double value = parMap.get(parName);
+            double err = parMap.get(parName + ".sd");
             dataValues.add(new ParValue(resNum, state, parName, value, err));
         });
         return dataValues;
